+++ CleverPush/CleverPush/Source/CPAppBannerViewController.m
@@ -306,7 +306,7 @@
     dispatch_async(dispatch_get_main_queue(), ^(void){
         [self fadeOut];
         [self jumpOut];
-        [[NSUserDefaults standardUserDefaults] setBool:false forKey:@"CleverPush_APP_BANNER_VISIBLE"];
+        [[NSUserDefaults standardUserDefaults] setBool:false forKey:CLEVERPUSH_APP_BANNER_VISIBLE_KEY];
         [[NSUserDefaults standardUserDefaults] synchronize];
         [self dismissViewControllerAnimated:NO completion:nil];
     });