+++ CleverPush/CleverPush/Source/CleverPushConstants.m
@@ -1,0 +1,9 @@
+//
+//  CleverPushConstants.m
+//  CleverPush
+//
+//  Created by Marius Gebhardt on 19/02/2022.
+//  Copyright © 2022 CleverPush. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>